tests/wpt/tests/css/css-overflow/column-creates-computed-style.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: ::column creates computed style</title>
<link rel="help" href="https://www.w3.org/TR/css-multicol-2/#column-pseudo">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#div::column { color: green }
</style>
<div id=div></div>
<script>
  test(() => {
  assert_equals(getComputedStyle(div, "::column").color, "rgb(0, 128, 0)");
  }, "::column creates computed style on its originating element");
</script>
