LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/column-creates-computed-style-expected.txt


FAIL ::column creates computed style on its originating element assert_equals: expected "rgb(0, 128, 0)" but got ""
